+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 -   `Text` can now rotate.
 -   `Button`. A button class that can be used to create a clickable area.
 -   `end()`. Quits the game.
+-   `pause()`. Pauses the game.
+-   `resume()`. Resumes the game.
 -   `penetration` and `normal` Manifold properties.
 
 ### Changed